feat(users): implement balance check before user deletion

+++ services/users/src/use-cases/delete-user-use-case.ts
@@ -3,6 +3,8 @@ import { providersEnum } from '../providers.enum.js';
 import { UserRepositoryInterface } from '../domain/user-repository-interface.js';
 import { type UseCaseInterface } from './use-case-interface.js';
 import { UserNotFoundException } from '../domain/error/user-not-found-exception.js';
+import { UserHashBalanceException } from '../domain/error/user-has-balance-exception.js';
+import { TransactionsServiceInterface } from '../domain/transactions-service-interface.js';
 
 interface DeleteUserUseCaseRequest {
   id: string;
@@ -17,7 +19,10 @@ export class DeleteUserUseCase
 {
   constructor(
     @Inject(providersEnum.UserRepository)
-    private readonly userRepository: UserRepositoryInterface
+    private readonly userRepository: UserRepositoryInterface,
+
+    @Inject(providersEnum.TransactionsService)
+    private readonly transactionsService: TransactionsServiceInterface
   ) {}
 
   async execute(request: DeleteUserUseCaseRequest): Promise<undefined> {
@@ -25,6 +30,12 @@ export class DeleteUserUseCase
 
     if (user === null) {
       throw new UserNotFoundException();
+    }
+
+    const balance = await this.transactionsService.getBalance(user.id);
+
+    if (balance !== 0) {
+      throw new UserHashBalanceException();
     }
 
     await this.userRepository.delete(user);
